Python-MQTT/requirements.txt
serial
paho-mqtt